fix protect events bug

+++ server/src/socket/deleteChat/index.ts
@@ -6,49 +6,66 @@ import { connectionQueryWrapper } from "../../libs";
 import ChatsService from "../../chats/chats.service";
 import SessionModel from "../../db/models/SessionModel";
 
+import { checkAuthSocket } from "../../libs";
+
 export const deleteChatHandler = (
 	io: Server,
 	socket: Socket,
 	usersSessions: Map<string, string>
 ) => {
-	socket.on("delete-chat", async (chatId: string) => {
-		try {
-			const myId = connectionQueryWrapper(socket.handshake.query.user);
+	socket.on(
+		"delete-chat",
+		async (
+			data: { chatId: string },
+			accessJwt: string,
+			cb: (err: { status: number; message: string }) => void
+		) => {
+			try {
+				const isAuth = checkAuthSocket(accessJwt, cb);
 
-			const chat = await ChatModel.findById(chatId);
-
-			if (!chat) {
-				io.emit("error-client", "Chat not found");
-				return;
-			}
-
-			const recipientId = chat.members.find((id) => id !== myId);
-
-			if (!recipientId) {
-				io.emit("error-client", "Recipient not found");
-				return;
-			}
-
-			await ChatsService.deleteChat(chatId);
-
-			const mySessions = await SessionModel.find({ user: myId });
-			const recipientSessions = await SessionModel.find({ user: recipientId });
-
-			mySessions.map((s) => {
-				const sessionSocketId = usersSessions.get(s._id.toString());
-				if (sessionSocketId) {
-					io.to(sessionSocketId).emit("delete-chat-client", { chatId });
+				if (!isAuth) {
+					return;
 				}
-			});
 
-			recipientSessions.map((s) => {
-				const sessionSocketId = usersSessions.get(s._id.toString());
-				if (sessionSocketId) {
-					io.to(sessionSocketId).emit("delete-chat-client", { chatId });
+				const { chatId } = data;
+
+				const myId = connectionQueryWrapper(socket.handshake.query.user);
+
+				const chat = await ChatModel.findById(chatId);
+
+				if (!chat) {
+					return cb({ status: 400, message: "Chat not found" });
 				}
-			});
-		} catch (e) {
-			io.emit("error-client", e);
+
+				const recipientId = chat.members.find((id) => id !== myId);
+
+				if (!recipientId) {
+					return cb({ status: 400, message: "Recipient not found" });
+				}
+
+				await ChatsService.deleteChat(chatId);
+
+				const mySessions = await SessionModel.find({ user: myId });
+				const recipientSessions = await SessionModel.find({
+					user: recipientId
+				});
+
+				mySessions.map((s) => {
+					const sessionSocketId = usersSessions.get(s._id.toString());
+					if (sessionSocketId) {
+						io.to(sessionSocketId).emit("delete-chat-client", { chatId });
+					}
+				});
+
+				recipientSessions.map((s) => {
+					const sessionSocketId = usersSessions.get(s._id.toString());
+					if (sessionSocketId) {
+						io.to(sessionSocketId).emit("delete-chat-client", { chatId });
+					}
+				});
+			} catch (e) {
+				cb({ status: 500, message: "Unexpected error" });
+			}
 		}
-	});
+	);
 };
